# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger documentation for task-API",
        "description": "Notes API service, main project -> https://github.com/OleGK4/task-api.git",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notes/{note}/tags": {
            "get": {
                "summary": "Display tags associated with the specified note",
                "operationId": "ae5c4e3d830ddbbbd1f6ff867bae99ce",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note tags paginated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteTagResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You do not own this note."
                    },
                    "404": {
                        "description": "Note not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Add a tag to the specified note",
                "operationId": "4d6f9ef04f1ede514539fbc4e117cddb",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note tag created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteTagResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You do not own this note."
                    },
                    "404": {
                        "description": "Note not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notes/{note}/tags/{tag}": {
            "get": {
                "summary": "Display the specified note tag",
                "operationId": "2d0eaf4c2fa9ee5ec988e590aa61b2f8",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "ID of the tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note tag retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteTagResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You do not own this note."
                    },
                    "404": {
                        "description": "Note tag or note not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Remove the specified note tag",
                "operationId": "9372b51ba1b478a254bfc0127e0c3612",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tag",
                        "in": "path",
                        "description": "ID of the tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note tag deleted successfully"
                    },
                    "403": {
                        "description": "You do not own this note."
                    },
                    "404": {
                        "description": "Note tag or note not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "NoteTagResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "tag": {
                        "$ref": "#/components/schemas/TagResource"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TagResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}